build-android-kernel: Remove "ZIP_BUILDER" variable
The concept of that variable was to make zipping optional, but... that now does
not make sense because if AnyKernel directory is defined that would mean
making a zip would be assumed to be wanted.

Also, at the current state of the script, the only thing this variable does
is something like "duplicate choice" or some sort of unneeded comfirmation.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>