build-android-kernel: Remove pointless quotes or curly braces
Even though using curly braces everywhere is considered a good bash programming
practice, it often isn't. In some cases, this practice can even make the code
needlessly hard to read/comprehend. Better use them only when needed, or when
readability is improved.

As for the quotes - they serve no purpose when used on variables or strings
that have no chance to contain whitespace. As unexpected as it is, they can do
more harm than good when some arguments of a command are contained in a
variable, and you accidentally or not, call the variable with quotes.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>