decrement_num: Remove dependency on directory for renamed files
The script will now use an own automatic temporary directory for the
decremented files. After work is done, it will automatically return all
files to their original directory.

The only disadvantage doing it this way, is that we need to move the
files again, doubling the time the script will have to run. Well, after
all, it all depends on the number of files. It barely takes a few
seconds to move a hundred files, so that is not of much concern.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>