spatch: Get the absolute path of a relative one
When the script is added to the path and called from a different
directory on a relative file path, $1 (file path) will not reflect the
absolute file path, hence using $1 as a file location in a command will
break the command, as it will not write to the actual file
location but rather a relative (inexistent) one.

To fix that, try to get the absolute path of the file if a relative one
is passed.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>